refactor(runtime-client): the package's source moves into `src`
The package kept its source at its own root -- four directories and
seven modules sitting beside `deno.jsonc` -- while `integration` and
`test` were already directories of their own. `src` joins them, so the
root now holds exactly the config and the three trees, and the layout
says which of them is the package.

Forty-six relative specifiers rewritten, each so that it resolves to the
module it resolved to before: the escaping `../../runner/src/...` imports
gain a level, the `test` tree's `../` become `../src/`, and everything
whose target moved with it is unchanged by construction. The two package
exports repoint at `src`, which is the whole of what other packages see
-- nothing outside this package imports it any other way, bar two files
in `shell`.

One of those two is a build config, not a test: `felt.config.ts` names
the worker entry by path. Verified by building the shell rather than by
type-checking it, since a bad entry resolves at build time and not
before -- `deno task build` emits `scripts/worker-runtime.js` from
`runtime-client/src/backends/web-worker/index.ts`. Its neighbouring
warning about `fetchBuildHash` is about the `out` path, which this does
not touch.

Eleven references in seven live documents follow the move.
`docs/history/` does not, by the rule that its documents are
point-in-time records: they describe a tree that had no `src`, and they
were right when written.

The run's inventory is identical -- 329 lines of `describe` and `it`
names, same order.
