+++ packages/runtime-client/deno.jsonc
@@ -7,7 +7,7 @@
     "integration": "LOG_LEVEL=warn deno test -A $INTEGRATION_TEST_FLAGS \"./integration/*.test.ts\""
   },
   "exports": {
-    ".": "./mod.ts",
-    "./transports/web-worker": "./client/transports/web-worker/transport-web-worker.ts"
+    ".": "./src/mod.ts",
+    "./transports/web-worker": "./src/client/transports/web-worker/transport-web-worker.ts"
   }
 }
