+++ packages/runtime-client/src/protocol/guards.ts
