+++ packages/runtime-client/src/shared/utils.ts
