+++ packages/runtime-client/test/client/connection.test.ts
@@ -9,8 +9,8 @@ import { getLogger } from "@commonfabric/utils/logger";
 import {
   consoleMessageFrom,
   RuntimeConnection,
-} from "../../client/connection.ts";
-import { EventEmitter } from "../../client/emitter.ts";
+} from "../../src/client/connection.ts";
+import { EventEmitter } from "../../src/client/emitter.ts";
 import {
   type ConsoleNotification,
   type InitializationData,
@@ -19,11 +19,11 @@ import {
   NotificationType,
   RequestType,
   RuntimeErrorCode,
-} from "../../protocol/mod.ts";
+} from "../../src/protocol/mod.ts";
 import type {
   RuntimeTransport,
   RuntimeTransportEvents,
-} from "../../client/transport.ts";
+} from "../../src/client/transport.ts";
 
 /**
  * Transport that records everything sent and auto-acknowledges every request
