+++ packages/runtime-client/test/schema-tags.test.ts
@@ -2,7 +2,7 @@ import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
 
 import type { JSONSchema } from "@commonfabric/api";
-import { tagsFromSchema } from "../schema-tags.ts";
+import { tagsFromSchema } from "../src/schema-tags.ts";
 
 describe("tagsFromSchema", () => {
   it("reads structured tags from the root", () => {
