+++ packages/shell/felt.config.ts
@@ -16,7 +16,7 @@ const config: Config = {
   entries: [
     { in: "src/index.ts", out: "scripts/index" },
     {
-      in: "../runtime-client/backends/web-worker/index.ts",
+      in: "../runtime-client/src/backends/web-worker/index.ts",
       // Changing this path requires a matching update in
       // packages/shell/src/lib/runtime.ts (fetchBuildHash).
       out: "scripts/worker-runtime",
